Custom Specie Name (#994)
<!--
This is a semi-strict format, you can add/remove sections as needed but
the order/format should be kept the same
Remove these comments before submitting
-->

<!--
Explain this PR in as much detail as applicable

Some example prompts to consider:
How might this affect the game? The codebase?
What might be some alternatives to this?
How/Who does this benefit/hurt [the game/codebase]?
-->

Add a Custom Specie Name for servers that wishes to allow players to
have custom names.
This simply add some variables to the Profile/Humanoid, meaning copying
the profile will keep the custom names.

By default its is disabled, for servers owners wishing to enable those
please edit the Specie Prototype and add:
```
customName: true
```

Enabling it will make it able for players to edit the specie name of the
selected specie, and will have the window enabled.

This is my first time fucking with Profile and Database, i fucking hate
it.

---

<!--
This is default collapsed, readers click to expand it and see all your
media
The PR media section can get very large at times, so this is a good way
to keep it clean
The title is written using HTML tags
The title must be within the <summary> tags or you won't see it
-->

<details><summary><h1>Media</h1></summary>
<p>

![image](https://github.com/user-attachments/assets/a6226fef-c015-45be-add9-ff8e3660d4bb)

![image](https://github.com/user-attachments/assets/cf440f13-5882-44fb-863a-4ba7d8dd1f68)

</p>
</details>

---

<!--
You can add an author after the `:cl:` to change the name that appears
in the changelog (ex: `:cl: Death`)
Leaving it blank will default to your GitHub display name
This includes all available types for the changelog
-->

:cl:
- add: Custom Species Names.

---------

Signed-off-by: FoxxoTrystan <45297731+FoxxoTrystan@users.noreply.github.com>
Signed-off-by: VMSolidus <evilexecutive@gmail.com>
Co-authored-by: Mnemotechnican <69920617+Mnemotechnician@users.noreply.github.com>
Co-authored-by: VMSolidus <evilexecutive@gmail.com>